+++ src/promptflow/promptflow/__init__.py
@@ -13,8 +13,8 @@ from promptflow._core.tracer import trace
 from promptflow._sdk._load_functions import load_flow, load_run
 
 from ._sdk._pf_client import PFClient
-from ._trace._start_trace import start_trace
 from ._version import VERSION
+from .tracing._start_trace import start_trace
 
 # backward compatibility
 log_flow_metric = log_metric
